backend/requirements.txt
fastapi
uvicorn[standard]
pandas
google-generativeai
python-multipart
